minor fixes

+++ dino_noise_experiments.py
@@ -547,61 +547,45 @@ def apply_tv(outputs_dir: Path, results_dir: Path, cfg: dict, cfg_name:str):
     
     # exit()
     
-    results_dict = OrderedDict()
-    with open(results_dir / "metrics.json", "r") as json_file:
-        results_dict = json.load(json_file, object_pairs_hook=OrderedDict)
-    if strategy['noise']['finetuning'][0]['noise_type'] == 'asymmetric':
-        alphas = tqdm(np.round(np.linspace(-2.05, -3.0, 20), 2))
-    else:
-        alphas = tqdm(np.round(np.linspace(-2.05, -3.0, 20), 2))
-    for alpha in alphas:
-        
-        model.load_state_dict(mix_weights, strict=False)
-        task_vectors['Average'].apply_to(model, scaling_coef=alpha, strict=False)
-        tv_test_results, _, _ = evaluate_model(model, dataset.get_test_dataloader(), gpu)
-        tv_train_results = eval_model_on_clean_noise_splits(model, None, dataset, gpu)
 
-        results_dict[alpha] = {'test_results': tv_test_results, 'train_results': tv_train_results}
-    with open(results_dir / 'metrics.json' , 'w') as json_file:
-        json.dump(results_dict, json_file, indent=4)
     
-    # results_dict = OrderedDict()
-    # if not results_dir.joinpath('metrics.json').exists():
+    results_dict = OrderedDict()
+    if not results_dir.joinpath('metrics.json').exists():
 
-    #     model.load_state_dict(mix_weights, strict=False)
-    #     mix_test_results, _, _ = evaluate_model(model, dataset.get_test_dataloader(), gpu)
-    #     mix_train_results = eval_model_on_clean_noise_splits(model, None, dataset, gpu)
+        model.load_state_dict(mix_weights, strict=False)
+        mix_test_results, _, _ = evaluate_model(model, dataset.get_test_dataloader(), gpu)
+        mix_train_results = eval_model_on_clean_noise_splits(model, None, dataset, gpu)
         
         
-    #     model.load_state_dict(gold_weights, strict=False)
-    #     gold_test_results, _, _ = evaluate_model(model, dataset.get_test_dataloader(), gpu)
-    #     gold_train_results = eval_model_on_clean_noise_splits(model, None, dataset, gpu)
+        model.load_state_dict(gold_weights, strict=False)
+        gold_test_results, _, _ = evaluate_model(model, dataset.get_test_dataloader(), gpu)
+        gold_train_results = eval_model_on_clean_noise_splits(model, None, dataset, gpu)
         
-    #     model.load_state_dict(ft_ho_clean_weights, strict=False)
-    #     ft_ho_test_results, _, _ = evaluate_model(model, dataset.get_test_dataloader(), gpu)
-    #     ft_ho_train_results = eval_model_on_clean_noise_splits(model, None, dataset, gpu)
+        model.load_state_dict(ft_ho_clean_weights, strict=False)
+        ft_ho_test_results, _, _ = evaluate_model(model, dataset.get_test_dataloader(), gpu)
+        ft_ho_train_results = eval_model_on_clean_noise_splits(model, None, dataset, gpu)
         
-    #     results_dict['Mix'] = {'test_results': mix_test_results, 'train_results': mix_train_results}
-    #     results_dict['Gold'] = {'test_results': gold_test_results, 'train_results': gold_train_results}
-    #     results_dict['FT HO Clean'] = {'test_results': ft_ho_test_results, 'train_results': ft_ho_train_results}
+        results_dict['Mix'] = {'test_results': mix_test_results, 'train_results': mix_train_results}
+        results_dict['Gold'] = {'test_results': gold_test_results, 'train_results': gold_train_results}
+        results_dict['FT HO Clean'] = {'test_results': ft_ho_test_results, 'train_results': ft_ho_train_results}
         
-    #     if strategy['noise']['finetuning'][0]['noise_type'] == 'asymmetric':
-    #         alphas = tqdm(np.round(np.linspace(-0.05, -2.0, 40), 2))
-    #     else:
-    #         alphas = tqdm(np.round(np.linspace(-0.05, -2.0, 40), 2))
-    #     for alpha in alphas:
+        if strategy['noise']['finetuning'][0]['noise_type'] == 'asymmetric':
+            alphas = tqdm(np.round(np.linspace(-0.05, -3.0, 60), 2))
+        else:
+            alphas = tqdm(np.round(np.linspace(-0.05, -3.0, 60), 2))
+        for alpha in alphas:
             
-    #         model.load_state_dict(mix_weights, strict=False)
-    #         task_vectors['Average'].apply_to(model, scaling_coef=alpha, strict=False)
-    #         tv_test_results, _, _ = evaluate_model(model, dataset.get_test_dataloader(), gpu)
-    #         tv_train_results = eval_model_on_clean_noise_splits(model, None, dataset, gpu)
+            model.load_state_dict(mix_weights, strict=False)
+            task_vectors['Average'].apply_to(model, scaling_coef=alpha, strict=False)
+            tv_test_results, _, _ = evaluate_model(model, dataset.get_test_dataloader(), gpu)
+            tv_train_results = eval_model_on_clean_noise_splits(model, None, dataset, gpu)
 
-    #         results_dict[alpha] = {'test_results': tv_test_results, 'train_results': tv_train_results}
-    #     with open(results_dir / 'metrics.json' , 'w') as json_file:
-    #         json.dump(results_dict, json_file, indent=4)
-    # else:
-    #     with open(results_dir / "metrics.json", "r") as json_file:
-    #         results_dict = json.load(json_file, object_pairs_hook=OrderedDict)
+            results_dict[alpha] = {'test_results': tv_test_results, 'train_results': tv_train_results}
+        with open(results_dir / 'metrics.json' , 'w') as json_file:
+            json.dump(results_dict, json_file, indent=4)
+    else:
+        with open(results_dir / "metrics.json", "r") as json_file:
+            results_dict = json.load(json_file, object_pairs_hook=OrderedDict)
             
             
     if 'alpha_KNN' not in results_dict:
